requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0 
asyncpg>=0.27.0
alembic>=1.7.0
mistralai>=0.0.7
openai>=1.0.0
pydantic-settings>=2.0.0